Rework how capture segments are handled a bit.
This change corrects for the fact that the block may receive
capture segment info while it is not currently writing a block.
Since capture segments apply to all samples going forward from
when they are received, this change correctly bootstraps the
first capture segment of any new recording with the old capture
segment data.